# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MSS Sample Stocking Platform API
  version: 1.0.0
  description: Contract for overview, demand collection, execution tracking and master data configuration.
servers:
  - url: http://localhost:8787/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /collection/plans:
    get:
      tags: [Collection]
      summary: List collection plans scoped to the current role
      parameters:
        - name: keyword
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/PlanStatus' }
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/RegionId'
      responses:
        '200':
          description: Collection plan list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CollectionPlan' }
    post:
      tags: [Collection]
      summary: Create a GTM collection plan
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, regionIds, deadline]
              properties:
                productId: { type: string }
                regionIds:
                  type: array
                  minItems: 1
                  items: { type: string }
                deadline: { type: string, format: date-time }
                note: { type: string, maxLength: 500 }
      responses:
        '201':
          description: Plan created in DRAFT state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /collection/plans/{planId}:
    get:
      tags: [Collection]
      summary: Get plan progress, summary and feedback
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /collection/plans/{planId}/release:
    post:
      tags: [Collection]
      summary: Release a plan to the MSS domain owner
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /collection/plans/{planId}/regions/{regionId}/draft:
    put:
      tags: [Collection]
      summary: Idempotently save a regional demand draft
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - $ref: '#/components/parameters/RegionIdPath'
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemandDraft'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /collection/plans/{planId}/regions/{regionId}/submit:
    post:
      tags: [Collection]
      summary: Submit regional demand to the MSS domain owner
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - $ref: '#/components/parameters/RegionIdPath'
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                version: { type: integer }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /collection/plans/{planId}/domain-feedback:
    post:
      tags: [Collection]
      summary: Formally hand the domain summary back to GTM
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmed, note]
              properties:
                confirmed: { type: boolean, enum: [true] }
                note: { type: string, minLength: 1, maxLength: 1000 }
                version: { type: integer }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /collection/plans/{planId}/regions/{regionId}/return:
    post:
      tags: [Collection]
      summary: Return a submitted regional snapshot for correction
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - $ref: '#/components/parameters/RegionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1 }
                version: { type: integer }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /collection/plans/{planId}/export:
    post:
      tags: [Collection]
      summary: Generate a production-planning export snapshot
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoleHeader:
      name: Authorization
      in: header
      required: true
      description: JWT bearer token. Retained as a reusable parameter for existing operation entries.
      schema: { type: string, example: 'Bearer eyJ...' }
    PlanId:
      name: planId
      in: path
      required: true
      schema: { type: string }
    ProductId:
      name: productId
      in: query
      schema: { type: string, default: all }
...
    RegionId:
      name: regionId
      in: query
      schema: { type: string }
    RegionIdPath:
      name: regionId
      in: path
      required: true
      schema: { type: string }
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Forbidden:
      description: Role or scope denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: State or version conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    Envelope:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: success }
        data: {}
        requestId: { type: string }
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            details:
              type: array
              items: { type: string }
    PlanStatus:
      type: string
      enum: [PRODUCT_DRAFT, READY_TO_RELEASE, COLLECTING, DOMAIN_REVIEW, GTM_CLOSURE, EXPORTED]
    CollectionPlan:
      type: object
      required: [id, productId, regionIds, submittedRegionIds, status, deadline, version]
      properties:
        id: { type: string }
        productId: { type: string }
        regionIds: { type: array, items: { type: string } }
        submittedRegionIds: { type: array, items: { type: string } }
        status: { $ref: '#/components/schemas/PlanStatus' }
        deadline: { type: string, format: date-time }
        demandTotal: { type: integer, minimum: 0 }
        note: { type: string }
        version: { type: integer, minimum: 1 }
    DemandDraft:
      type: object
      required: [items]
      properties:
        version: { type: integer }
        items:
          type: array
          items:
            type: object
            required: [productItemKey, officeId, quantity]
            properties:
              productItemKey: { type: string, description: SKU id or product-level provisional key }
              officeId: { type: string, description: Representative office in the selected region }
              quantity: { type: integer, minimum: 0 }
              basis: { type: string }
              plannedUseDate: { type: string, format: date }
              note: { type: string, maxLength: 500 }
...